requirements.txt
Django>=4.2,<5
psycopg2-binary>=2.9
djangorestframework>=3.15
python-dotenv>=1.0
